etgenai/requirements.txt
# ── Web framework ────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0

# ── Agentic framework ─────────────────────────────────────────────
langgraph==0.3.0
langchain==0.2.0

# ── LLM client (Ollama via HTTP) ──────────────────────────────────
httpx==0.27.0

# ── Data validation (Python 3.13 compatible) ──────────────────────
pydantic>=2.10.0
pydantic-settings>=2.5.0

# ── Async + scheduling ────────────────────────────────────────────
python-multipart==0.0.9

# ── Env management ────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Dev / testing ─────────────────────────────────────────────────
pytest==8.3.0
pytest-asyncio==0.23.8
